requirements.txt
bandit==1.6.2
black==20.8b1
numpy==1.20.3
Pillow==6.2.1
Pint==0.17
pylint==2.8.2
pytest==6.2.4
pytest-cov==2.12.0
PyYAML==5.2
wily==1.19.0
